+++ tests/unit/driver/core/test_core_drivers.py
@@ -53,3 +53,37 @@ def test_juniper_driver_init_telnet():
 def test_nxos_driver_init_telnet():
     conn = NXOSDriver(host="myhost", transport="telnet")
     assert conn.transport.username_prompt == "login:"
+
+
+@pytest.mark.parametrize(
+    "session_data",
+    [
+        (NXOSDriver, "mysession", r"^[a-z0-9.\-@/:]{1,32}\(config\-s[a-z0-9.\-@/:]{0,32}\)#\s?$"),
+        (
+            EOSDriver,
+            "mysession",
+            r"^[a-z0-9.\-@/:]{1,32}\(config\-s\-mysess[a-z0-9_.\-@/:]{0,32}\)#\s?$",
+        ),
+    ],
+    ids=["nxos", "eos"],
+)
+def test_driver_register_configuration_session(session_data):
+    driver = session_data[0]
+    session_name = session_data[1]
+    session_pattern = session_data[2]
+    conn = driver(host="myhost")
+    conn.register_configuration_session(session_name=session_name)
+    assert conn.privilege_levels[f"configuration_session_{session_name}"].pattern == session_pattern
+
+
+@pytest.mark.parametrize(
+    "driver", [NXOSDriver, EOSDriver,], ids=["nxos", "eos"],
+)
+def test_driver_register_configuration_session_invalid_name(driver):
+    conn = driver(host="myhost")
+    with pytest.raises(ValueError) as exc:
+        conn.register_configuration_session(session_name="1234")
+    assert (
+        str(exc.value)
+        == "Scrapli requires configuration session names to be valid python identifiers, provided session_name `1234` is not a valid identifier"
+    )